+++ Manager/FeedbackManager.php
@@ -6,6 +6,8 @@ use Doctrine\ORM\EntityManager;
 
 use Developtech\AgilityBundle\Model\ProjectModel;
 
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+
 class FeedbackManager {
     /** @var EntityManager **/
     protected $em;
@@ -27,5 +29,16 @@ class FeedbackManager {
      */
     public function getProjectFeedbacks(ProjectModel $project) {
         return $this->em->getRepository($this->feedbackClass)->findByProject($project);
+    }
+
+    /**
+     * @param integer $id
+     * @return FeedbackModel
+     */
+    public function getFeedback($id) {
+        if(($feedback = $this->em->getRepository($this->feedbackClass)->find($id)) === null) {
+            throw new NotFoundHttpException('Feedback not found');
+        }
+        return $feedback;
     }
 }