+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+### Added
+* Feedback model
+* Project feedbacks list service
+
 ## [0.1.0] - 2016-07-13
 ### Added
 * Project model