+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 * Feedback model
 * Project feedbacks list service
+* Get feedback service
 
 ## [0.1.0] - 2016-07-13
 ### Added