+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Feedback model
 * Project feedbacks list service
 * Get feedback service
+* Feedback creation service
 
 ## [0.1.0] - 2016-07-13
 ### Added